Alright. I found a lot of errors in some of the existing base code.
The EpisodeTests were not working correctly. I fixed this and found that
much of what appeared to be working was not actually working.
In particular, this is true of the ElementalEpisodes copmarisons. It turns
out that the equalSensors() method was not working. This was due to the fact
that the Hashtable comparison was not being used properly.

In light of this I added a class called WMESet. This is just a holder for the
hashtable we were using before, but I was able to use it define more precise
comparison routines for different WMESets.
I also modified the toString method for it so that the set is always 'stringed'
in alphabetical order with respect to the keys.

I wrote a junit test file for this new class. I also slightly modified the
EpisodeTest file so that it is actually making useful comparisons.