I changed update() to end sequences when the agent reaches a goal (in addition to ending them when the agent executes an indeterminate action.  *I think* this repaired a bug I was seeing with the agent's ability to find plans.
I changed all 'private' methods in Ziggurat.java to 'protected' for
two reasons: a) we may want to create subclasses with variant behavior
and b) I want the javadoc comments to show up.

I renamed the findInterimStart methods to findOrientation.  I think
that's a marginally better name and less confusing.