Consider this log snippet from the FlipSystem environment:
------------------

    Choosing command from plan:
        {
          Level 0: {1U---->0*, 0R---->1}
        }
    Enter: considerReplacement
        Enter: findBestReplacement
            searching for replacement at level 0
            best existing repl:
                { 1U---->0, 0R---->1 }==>1U---->0
        Exit: findBestReplacement
        Enter: makeNewReplacement
            Attempting to construct a new replacement for these two actions:
                1U---->0
                0R---->1
            Considering this action for the replacement: 0R---->1
            LHS sensors don't match, try a different candidate
            Considering this action for the replacement: 1L---->1
            RHS sensors don't match, try a different candidate
            Considering this action for the replacement: 1U---->0
            replacement already exists (duplicate)
            Considering this action for the replacement: 0U---->0
            LHS sensors don't match, try a different candidate
            No new replacement could be constructed from existing actions.

------------------

Q. What's so interesting about this?
A. This message never appears:
	Considering this action for the replacement: 1R---->1

If it had been considered it would be the perfect replacement.  Why was it never considered?  *Because the agent never experienced this action.  Ever.*

Currently the agent only considers random actions when it has to so it can potentially miss a critical action that is necessary to reach an optimal behavior.  Perhaps it should always consider a semi-random command when that semi-random command will yield a new action?

Issue:  The agent can't be sure whether a particular command will result in a new action since the outcome of the command is not known.  For now, let's punt on this issue and stick with overriding a plan only when the agent is sure a particular command will result in a new action.  In the long term we may want a way to deal for the agent to estimate the likelihood of a new outcome and decide based upon that.

Issue:  As the number of available actions and/or sensings increases, this approach will keep the agent from completing a plan an exponentially longer and longer time.  I'm punting on this issue for now as well.